+++ tests/HelperCacheTrait/HelperCacheRuntimeTest.php
@@ -20,21 +20,18 @@ final class HelperCacheRuntimeTest extends TestCase
         $mixed = Helper::cacheRuntime('keyA', fn () => 1);
         $mixed = Helper::cacheRuntimeGet('keyA');
         $this->assertTrue($mixed === 1);
-
         $mixed = Helper::cacheRuntime('keyA', fn () => 0);
         $this->assertTrue($mixed === 1);
 
         $mixed = Helper::cacheRuntime('keyB', fn () => '2');
         $mixed = Helper::cacheRuntimeGet('keyB');
         $this->assertTrue($mixed === '2');
-
         $mixed = Helper::cacheRuntime('keyB', fn () => 0);
         $this->assertTrue($mixed === '2');
 
         $mixed = Helper::cacheRuntime('keyC', fn () => ['3']);
         $mixed = Helper::cacheRuntimeGet('keyC');
         $this->assertTrue($mixed === ['3']);
-
         $mixed = Helper::cacheRuntime('keyC', fn () => 0);
         $this->assertTrue($mixed === ['3']);
     }